requirements.txt
 # requirements.txt
# Dependências existentes
gunicorn==21.2.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
typing_extensions==4.14.0
Werkzeug==3.1.3
pymongo==4.6.1
python-dotenv==1.0.0
redis==5.0.1
firebase-admin==6.4.0
cloudinary==1.39.0
whitenoise==6.10.0

# Dependências para processamento de documentos
reportlab==4.0.9
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.1
pandas==2.2.2
openpyxl==3.1.2
opencv-python==4.8.1.78
python-dateutil==2.8.2
regex==2023.12.25
psutil>=5.9.0
requests>=2.28.0

numpy==1.26.4

# Dependências adicionais necessárias
xlrd==2.0.1
lxml==4.9.4
beautifulsoup4==4.12.2
gevent

# ===== NOVAS DEPENDÊNCIAS JWT =====
PyJWT==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

#===== DEPENDENCIAS NLP =====



# ... todas as suas dependências existentes ...

# Adicionar estas linhas no final
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio
transformers
accelerate



